Plotting options
Added an option to bivectors along their tip, tail, or midpoint, where
the default option is now the midpoint instead of the tail. The "tip" is
just the top vertex, as measure from the tails of the spanning vectors
upward. It looks so much better now placing bivectors along the
midpoint, however, that I can't imagine why you would want to use the
other two options anyway.