Correct Orientation for Bivectors
BivectorPlot3D now performs a check to ensure that the blade will be in
the correct orientation before drawing the bivector. Since there is no
guarantee that the spanning vectors generated by NullSpace will be given
in the correct orientation, it was consequently not guaranteed that the
rotation indicated by the curved arrow drawn on the blade was correct or
not.

Additionally, the previous commit has been refactored a bit to be more
Mathematica-y, which was helpful for integrating it with the current
change, and certainly will also so for future changes.