Fixed bug regarding singularities from NullSpace
BivectorPlot3D now performs a check for any errors arising from
singularities in the frame provided by NullSpace, and refuses to draw
the bivector in that case. BivectorPlot3D also quiets certain expected
error messages that will arise in this case.

This also addresses temporarily an issue where the Poisson structures we
are interested in were not not being sampled evenly, as I had to
introduce a slightly adjusted step size in the Manipulate command to
"skip over" the singularities that were causing problems in that case.
It remains to be seen, however, whether this will be sufficient for even
sampling in other cases, or what other updates might be necessesary to
always ensure even sampling.